requirements.txt
streamlit==1.25.0
openai==1.12.0
python-dotenv
fastapi==0.85.1
pydantic==1.10.14
uvicorn==0.27.1